feat&refactor: add cache for pointUpdates put out single check point details
