+++ src/models/pointUpdates.ts
@@ -1,0 +1,4 @@
+export interface IPointUpdates {
+    url: string;
+    data: string;
+}
