+++ src/parser/pointUpdates.ts
@@ -2,8 +2,7 @@ import { executeRegex } from '../utils/sentry';
 
 const lastPointDateRegex = /([0-9]{2}\.[0-9]{2}\.[0-9]{4} )/m;
 
-export default function parsePointUpdates(xhr: string): string {
-    // когда дат изменения нет, запрос возвращает один пробел (но это не точно)
-    if (xhr == "" || xhr.length < 10) return "empty"; 
-    return executeRegex(lastPointDateRegex, xhr)[0];
+export default function parsePointUpdates(data: string): string {
+    if (data.length == 1) return null; 
+    return executeRegex(lastPointDateRegex, data)[0];
 }
